fix(firefox): skip symlink if exists

+++ firefox/install.sh
@@ -12,6 +12,8 @@ printf "%s: Installing user.js\n" "$0"
 profile=$(grep -Po -m1 'Default=\K.*' <~/.mozilla/firefox/profiles.ini)
 if [ -d "$XDG_CONFIG_HOME"/firefox ]; then
 	for item in "$XDG_CONFIG_HOME"/firefox/*; do
-		ln -vst ~/.mozilla/firefox/"$profile" "$item"
+		if [ ! -L ~/.mozilla/firefox/"$profile"/"${item##*/}" ]; then
+			ln -vst ~/.mozilla/firefox/"$profile" "$item"
+		fi
 	done
 fi
